# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexora AI Location Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /places/search:
    get:
      operationId: searchPlaces
      tags: [places]
      summary: Search for places
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of matching places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
  /places/{placeId}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get place details
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceDetail"
        "404":
          description: Not found
  /places/{placeId}/personality:
    get:
      operationId: getPlacePersonality
      tags: [places]
      summary: Get AI-generated place personality and area DNA
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Place personality profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacePersonality"
...
  /timelines/{placeId}:
    get:
      operationId: readPlaceTimeline
      tags: [timelines]
      summary: Get AI-generated timeline for a place
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            description: ISO date string, defaults to today
        - name: mode
          in: query
          schema:
            type: string
            enum: [today, yesterday, last_week, last_month, custom, festival, holiday]
            default: today
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeline"
  /timelines/{placeId}/replay:
    get:
      operationId: readTimelineReplay
      tags: [timelines]
      summary: Get time-sliced data for timeline replay animation
      parameters:
        - name: placeId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hourly replay data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplaySlice"
...
  /saved-places:
    get:
      operationId: listSavedPlaces
      tags: [saved]
      summary: List saved places for current session
      responses:
        "200":
          description: Saved places list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedPlace"
    post:
      operationId: savePlace
      tags: [saved]
      summary: Save a place
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedPlaceInput"
      responses:
        "201":
          description: Saved place created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPlace"
  /saved-places/{id}:
    delete:
      operationId: deleteSavedPlace
      tags: [saved]
      summary: Remove a saved place
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Place:
      type: object
      required: [id, name, country, lat, lng]
      properties:
        id: { type: string }
        name: { type: string }
        country: { type: string }
        state: { type: ["string", "null"] }
        lat: { type: number }
        lng: { type: number }
        placeType: { type: string }
        population: { type: ["integer", "null"] }
        timezone: { type: ["string", "null"] }
    PlaceDetail:
      type: object
      required: [id, name, country, lat, lng, description]
      properties:
        id: { type: string }
        name: { type: string }
        country: { type: string }
        state: { type: ["string", "null"] }
        lat: { type: number }
        lng: { type: number }
        placeType: { type: string }
        population: { type: ["integer", "null"] }
        timezone: { type: ["string", "null"] }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
    PlacePersonality:
      type: object
      required: [placeId, placeName, ratings, areaDna, summary]
      properties:
        placeId: { type: string }
        placeName: { type: string }
        ratings:
          type: object
          properties:
            energy: { type: number }
            tourism: { type: number }
            shopping: { type: number }
            nightlife: { type: number }
            education: { type: number }
            business: { type: number }
        areaDna:
          type: object
          properties:
            food: { type: number }
            business: { type: number }
            culture: { type: number }
            shopping: { type: number }
            nature: { type: number }
            entertainment: { type: number }
            education: { type: number }
            healthcare: { type: number }
        summary: { type: string }
        generatedAt: { type: string }
...
    Timeline:
      type: object
      required: [placeId, date, entries, narrative]
      properties:
        placeId: { type: string }
        date: { type: string }
        mode: { type: string }
        narrative: { type: string }
        entries:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEntry"
    TimelineEntry:
      type: object
      required: [time, title, description, category]
      properties:
        time: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string, enum: [traffic, crowd, restaurant, transit, event, weather, safety, business] }
        intensity: { type: string, enum: [low, medium, high] }
        icon: { type: string }
    ReplaySlice:
      type: object
      required: [hour, label]
      properties:
        hour: { type: integer }
        label: { type: string }
        traffic: { type: number }
        crowd: { type: number }
        noise: { type: number }
        activity: { type: number }
        description: { type: string }
...
    SavedPlace:
      type: object
      required: [id, placeId, placeName, lat, lng, createdAt]
      properties:
        id: { type: integer }
        placeId: { type: string }
        placeName: { type: string }
        country: { type: string }
        lat: { type: number }
        lng: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    SavedPlaceInput:
      type: object
      required: [placeId, placeName, lat, lng]
      properties:
        placeId: { type: string }
        placeName: { type: string }
        country: { type: string }
        lat: { type: number }
        lng: { type: number }
        notes: { type: string }
...