# lib/api-zod/src/generated/types/placeDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora AI Location Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceDetail {
  id: string;
  name: string;
  country: string;
  /** @nullable */
  state?: string | null;
  lat: number;
  lng: number;
  placeType?: string;
  /** @nullable */
  population?: number | null;
  /** @nullable */
  timezone?: string | null;
  description: string;
  tags?: string[];
  createdAt?: string;
}