# lib/api-zod/src/generated/types/placePersonality.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora AI Location Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PlacePersonalityAreaDna } from './placePersonalityAreaDna';
import type { PlacePersonalityRatings } from './placePersonalityRatings';

export interface PlacePersonality {
  placeId: string;
  placeName: string;
  ratings: PlacePersonalityRatings;
  areaDna: PlacePersonalityAreaDna;
  summary: string;
  generatedAt?: string;
}