# lib/api-zod/src/generated/types/timeline.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora AI Location Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEntry } from './timelineEntry';

export interface Timeline {
  placeId: string;
  date: string;
  mode?: string;
  narrative: string;
  entries: TimelineEntry[];
}